Launch pwa with appuser model id
