+++ src/modules/Workspaces/WorkspacesLauncher/AppLauncher.cpp
@@ -165,16 +165,41 @@ namespace AppLauncher
 
         if (!launched && !app.pwaAppId.empty())
         {
-            std::filesystem::path appPath(app.path);
-            if (appPath.filename() == NonLocalizable::EdgeFilename)
+            int version = 0;
+
+            if (app.version != L"")
             {
-                appPathFinal = appPath.parent_path() / NonLocalizable::EdgePwaFilename;
-                commandLineArgsFinal = NonLocalizable::PwaCommandLineAddition + app.pwaAppId + L" " + app.commandLineArgs;
+                version = std::stoi(app.version);
             }
-            if (appPath.filename() == NonLocalizable::ChromeFilename)
+
+            if (version >= 1)
             {
-                appPathFinal = appPath.parent_path() / NonLocalizable::ChromePwaFilename;
-                commandLineArgsFinal = NonLocalizable::PwaCommandLineAddition + app.pwaAppId + L" " + app.commandLineArgs;
+                // Should launch directaly from appUserModelId first.
+                auto res = LaunchApp(L"shell:AppsFolder\\" + app.appUserModelId, app.commandLineArgs, app.isElevated);
+                if (res.isOk())
+                {
+                    launched = true;
+                    Logger::trace("Successfully launched pwa apps through appusermodelid");
+                }
+                else
+                {
+                    launchErrors.push_back({ app.appUserModelId, res.error() });
+                }
+            }
+
+            if (!launched)
+            {
+                std::filesystem::path appPath(app.path);
+                if (appPath.filename() == NonLocalizable::EdgeFilename)
+                {
+                    appPathFinal = appPath.parent_path() / NonLocalizable::EdgePwaFilename;
+                    commandLineArgsFinal = NonLocalizable::PwaCommandLineAddition + app.pwaAppId + L" " + app.commandLineArgs;
+                }
+                if (appPath.filename() == NonLocalizable::ChromeFilename)
+                {
+                    appPathFinal = appPath.parent_path() / NonLocalizable::ChromePwaFilename;
+                    commandLineArgsFinal = NonLocalizable::PwaCommandLineAddition + app.pwaAppId + L" " + app.commandLineArgs;
+                }
             }
         }
